Set useFilterManager to handle its own subscription Removed subscription handling from useEsSearch
